+++ src/wallet/constants/WalletName.ts
@@ -5,6 +5,7 @@ export const WalletName = {
   STATION: "station",
   KEPLR: "keplr",
   LEAP: "leap",
+  COMPASS: "compass",
   COSMOSTATION: "cosmostation",
   METAMASK_INJECTIVE: "metamask-injective",
 } as const;