+++ src/wallet/index.ts
@@ -12,6 +12,7 @@ export {
   type ChainInfo,
   type EventCallback,
 } from "./wallets/WalletController";
+export { CompassController } from "./wallets/compass/CompassController";
 export { CosmostationController } from "./wallets/cosmostation/CosmostationController";
 export { KeplrController } from "./wallets/keplr/KeplrController";
 export { LeapController } from "./wallets/leap/LeapController";