+++ src/wallet/wallets/compass/CompassExtension.ts
@@ -1,0 +1,4 @@
+import { LeapExtension } from "../leap/LeapExtension";
+
+// Compass's API is similar to Leap.
+export const CompassExtension = LeapExtension;