+++ src/wallet/wallets/compass/types.ts
@@ -1,0 +1,8 @@
+import { Leap } from "../leap/types";
+
+// Type is similar to Leap
+export type Compass = Leap;
+
+export type Window = {
+  compass?: Compass | undefined;
+};