+++ src/wallet/wallets/window.d.ts
@@ -1,5 +1,6 @@
 import { Window as KeplrWindow } from "cosmes/registry";
 
+import { Window as CompassWindow } from "./compass/types";
 import { Window as CosmostationWindow } from "./cosmostation/types";
 import { Window as LeapWindow } from "./leap/types";
 import { Window as EthereumWindow } from "./metamask-injective/types";
@@ -11,5 +12,6 @@ declare global {
       CosmostationWindow,
       StationWindow,
       LeapWindow,
+      CompassWindow,
       EthereumWindow {}
 }