+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # Changelog
 
+## `v0.0.48`
+
+### Features
+
+- Added Compass wallet support to execute txs via Compass extension on Sei network (*does not support wallet connect yet*)
+
 ## `v0.0.47`
 
 ### Features
 
-- Added `MetamaskInjective` wallet to execute txs via MetaMask extension on Injective network (*does not support wallet connect yet*)
+- Added MetaMask wallet support to execute txs via MetaMask extension on Injective network (*does not support wallet connect yet*)
 
 ### Improvements